<<<<<<< HEAD
\name{secrBVN-package}
\alias{secrBVN-package}
\alias{secrBVN}

\docType{package}

\title{ Bivariate Normal Home Range SECR }

\description{

  Functions to simulate and plot random bivariate normal home ranges,
  extending relevant capability in secr. Previously named `bvn' 
  
}

\details{
\tabular{ll}{
Package: \tab secrBVN\cr
Type: \tab Package\cr
Version: \tab 1.2.2\cr
Date: \tab 2018-07-26\cr
License: \tab GNU General Public License Version 2 or later\cr
}

\tabular{ll}{
\code{\link{plotpopn.bvn}} \tab \cr
\code{\link{simcapt.bvn}} \tab \cr
\code{\link{simpopn.bvn}} \tab \cr
\code{\link{runEllipseSim}} \tab \cr
}

}

\author{
Murray Efford \email{murray.efford@otago.ac.nz}
}

\keyword{ package }
=======
\name{secrBVN-package}
\alias{secrBVN-package}
\alias{secrBVN}

\docType{package}

\title{ Bivariate Normal Home Range SECR }

\description{

  Functions to simulate and plot random bivariate normal home ranges,
  extending relevant capability in secr. Previously named `bvn' 
  
}

\details{
\tabular{ll}{
Package: \tab secrBVN\cr
Type: \tab Package\cr
Version: \tab 1.2.2\cr
Date: \tab 2018-07-26\cr
License: \tab GNU General Public License Version 2 or later\cr
}

\tabular{ll}{
\code{\link{plotpopn.bvn}} \tab \cr
\code{\link{simcapt.bvn}} \tab \cr
\code{\link{simpopn.bvn}} \tab \cr
\code{\link{runEllipseSim}} \tab \cr
}

}

\author{
Murray Efford \email{murray.efford@otago.ac.nz}
}

\keyword{ package }
>>>>>>> 67bfb5b162dab3c4406b76b4faf48eabee35533f
